Fix logic for setting keyboard repeat rate and delay
- Repeat rate scale is not linear, it's something like an inverse exponential that depends on two inputs, so inverting that function isn't easy
- To solve this, we'll have the user look up which byte they need from a table
- We may as well let the user look up the delay byte they need too, since the table for delay is right underneath the one for repeat rate
- As a bonus, we can remove libm as a dependency

+++ src/error.rs
@@ -12,8 +12,8 @@ pub enum KeyboardError {
     Resend,
     KeyDetectionError,
     InvalidResponse(u8),
-    InvalidTypematicFrequency(f64),
-    InvalidTypematicDelay(u16),
+    InvalidTypematicFrequency(u8),
+    InvalidTypematicDelay(u8),
     ControllerError(ControllerError),
 }
